second commit

+++ order_app/models.py
@@ -22,6 +22,6 @@ class Order(models.Model):
     customer_id=models.ForeignKey(Customer, on_delete = models.CASCADE)
     product_id=models.ForeignKey(Product, on_delete = models.CASCADE)
     price =models.DecimalField(max_digits=10, decimal_places=2)
-    qty=models.DecimalField(max_digits=10, decimal_places=2)
+    qty=models.IntegerField()
     total_price=models.DecimalField(max_digits=10, decimal_places=2)
     created_at = models.DateTimeField(auto_now_add=True)
